tests/wpt/tests/svg/animations/animate-from-to-keyTimes.html

<!doctype html>
<html>
<meta charset="utf-8">
<title>Tests discrete from-to-keyTimes animations</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/SVGAnimationTestCase-testharness.js"></script>

<svg>
</svg>

<script>
var rootSVGElement = document.querySelector("svg");
var epsilon = 1.0;

// Setup test document
var rect = createSVGElement("rect");
rect.setAttribute("id", "rect");
rect.setAttribute("x", "100");
rect.setAttribute("width", "100");
rect.setAttribute("height", "100");
rect.setAttribute("fill", "green");
rect.setAttribute("onclick", "executeTest()");

var animate = createSVGElement("animate");
animate.setAttribute("id", "animation");
animate.setAttribute("attributeName", "x");
animate.setAttribute("to", "200");
animate.setAttribute("dur", "4s");
animate.setAttribute("keyTimes", "0;0.25");
animate.setAttribute("calcMode", "discrete");
animate.setAttribute("fill", "freeze");
rect.appendChild(animate);
rootSVGElement.appendChild(rect);

// Setup animation test
function sample1() {
    assert_equals(rect.x.animVal.value, 100);
    assert_equals(rect.x.baseVal.value, 100);
}

function sample2() {
    assert_equals(rect.x.animVal.value, 200);
    assert_equals(rect.x.baseVal.value, 100);
}

smil_async_test((t) => {
    const expectedValues = [
        // [animationId, time, sampleCallback]
        ["animation", 0.5, sample1],
        ["animation", 1.5, sample2],
        ["animation", 2.5, sample2]
    ];

    runAnimationTest(t, expectedValues);
});

window.clickX = 150;

</script>
